+++ distributeblocks/Wallet.java
@@ -1,9 +1,10 @@
-package DistributeBlocks;
+package distributeblocks;
 
 import java.security.*;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
+import distributeblocks.crypto.*;
 
 public class Wallet {
