+++ distributeblocks/util/Validator.java
@@ -1,6 +1,8 @@
-package DistributeBlocks;
+package distributeblocks.util;
 
 import java.util.LinkedList;
+import distributeblocks.*;
+import distributeblocks.crypto.*;
 
 public class Validator
 {
